Move test to simapp?
